Added terrain, grass, terrain textures, etc...
Big terrain update -> Added terrain and dirt pathways, grass, and mountains encompassing the area. Also added a 'ranger hut' which is a structure. Need to find a 'drug' sprite or something that looks like cocaine for the powerup. Also need to add script for powerup.

Added UI component which allows the player to see their health and cocaine levels.