Animations, trees, grass, and areas
Added bear animations for running (still need to get attack and jump working). They are triggers which are different than booleans for animations.
Added grass and trees.
Added 'Areas' which will later serve as hooks for NPCs to walk to. Areas also have cocaine in them, allowing the bear to sustain itself. I need to find a way to standardize the cocaine bag, but the script to control it is working and its adds 50 cocaine to the bear's cocaine bar.